airflow/requirements.txt
psycopg2-binary
aiohttp
beautifulsoup4
lxml